Clarify email messages
This tries to clarify the email messaging to avoid alarming users.

I also hit the file with the `black` Python formatter.  (Considering adding
this to CI.)

I left an override in place to email paul.hirsch@gsa.gov if you want to
apply this to test.  Once things look good we can pull the safety and
ship it.