Add missing certs for the US Courts (#362)
We received a support request about PIV issues at US Courts. I used the rake task to hunt for the missing certs. This commit includes those certs along with some changes to the rake task so it doesn't break when you try to use it.

This commit is missing one certificate because it requires a change to the allowed hosts in the OBProxy. This is necessary to fetch CRLs for the CA. That certificate is added to #363 which can be merged once that infrastructure change is applied.